Add `find_child` method to `FortranBase`
